// apps/stm-g/src/modules-features/admin/MELevelManagement/interfaces.ts
export interface I_Level {
    id?: number
    code?: string;
    name?: string;
    programCode?: string;
    levelDescription?: string;
    levelOrder?: number;
    totalLessons?: number;
    totalHours?: number;
    suggestedFeeNote?: string;
    finalTestCode?: string;
    isActive?: boolean;
}